Publish offer method added.

+++ EbaySharp/Entities/Sell/Inventory/Offer/Tax.cs
@@ -4,6 +4,6 @@
     {
         public bool ApplyTax { get; set; }
         public string ThirdPartyTaxCategory { get; set; }
-        public int VatPercentage { get; set; }
+        public float VatPercentage { get; set; }
     }
 }
